# Optimisations for wide datasets and deep trees