@RAMitchell 2024-10-18
The decrease in performance for covtype can be traced to features_per_warp 64->32. I guess this is because covtype has few bins per feature on average.