@RAMitchell 2024-10-18
On the benchmark dataset
`legate --fbmem 29000 --sysmem 30000 --eager-alloc-percentage 20 --gpus 1 scaling.py --nrows 10000000 --ncols 100 --output test.csv --repeats 3 --model_types tree`
main: 17.771741s
this PR: 19.313279s

I will look into this further before merging.